complete refactor with message struct
