+++ processors/push.go
@@ -2,9 +2,8 @@ package processors
 
 import (
 	"log"
-	"net/http"
 	"github.com/ZacharyGroff/GitHooks/config"
-	"github.com/ZacharyGroff/GitHooks/endpoint"
+	"github.com/ZacharyGroff/GitHooks/models"
 )
 
 type PushProcessor struct {
@@ -16,7 +15,7 @@ func NewPushProcessor(deployer *Deployer, config *config.Config) *PushProcessor 
 	return &PushProcessor{deployer, config}
 }
 
-func (pushProcessor PushProcessor) HandleRequest(message *main.Message) {
+func (pushProcessor PushProcessor) HandleMessage(message *models.Message) {
 	output := pushProcessor.Deployer.Deploy()
 	log.Printf("Successfully deployed script with output:\n%s\n", output)
 }
