+++ wire.go
@@ -7,11 +7,12 @@ import (
 	"github.com/ZacharyGroff/GitHooks/config"
 	"github.com/ZacharyGroff/GitHooks/processors"
 	"github.com/ZacharyGroff/GitHooks/validation"
+	"github.com/ZacharyGroff/GitHooks/endpoint"
 )
 
-func InitializeEndpoint() Endpoint {
-	wire.Build(NewEndpoint, processors.NewPushProcessor, processors.NewDeployer, 
+func InitializeEndpoint() endpoint.Endpoint {
+	wire.Build(endpoint.NewEndpoint, processors.NewPushProcessor, processors.NewDeployer, 
 		config.NewConfig, validation.NewValidator)
 
-	return Endpoint{}
+	return endpoint.Endpoint{}
 }
