+++ wire_gen.go
@@ -7,17 +7,18 @@ package main
 
 import (
 	"github.com/ZacharyGroff/GitHooks/config"
+	"github.com/ZacharyGroff/GitHooks/endpoint"
 	"github.com/ZacharyGroff/GitHooks/processors"
 	"github.com/ZacharyGroff/GitHooks/validation"
 )
 
 // Injectors from wire.go:
 
-func InitializeEndpoint() Endpoint {
+func InitializeEndpoint() endpoint.Endpoint {
 	configConfig := config.NewConfig()
 	deployer := processors.NewDeployer(configConfig)
 	pushProcessor := processors.NewPushProcessor(deployer, configConfig)
 	validator := validation.NewValidator(configConfig)
-	endpoint := NewEndpoint(configConfig, pushProcessor, validator)
-	return endpoint
+	endpointEndpoint := endpoint.NewEndpoint(configConfig, pushProcessor, validator)
+	return endpointEndpoint
 }
